tests/wpt/web-platform-tests/css/css-tables/percent-width-ignored-001.tentative.html

<!DOCTYPE html>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='/resources/check-layout-th.js'></script>
<link rel="author" title="David Grogan" href="dgrogan@chromium.org">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/3336">
<meta name="assert" content="A cell's percent width is ignored when its table is nested in another cell" />
No red should show. Yellow and blue rectangles are in proportion to their
contents' intrinsic widths, not affected by yellow's percent-width bloating the
inner table.
<table id="outerTable" style="width: 300px" cellspacing="0" cellpadding="0">
    <td style="background:red;">
        <table cellspacing="0" cellpadding="0">
            <td style="width:1%; background:yellow;" data-expected-width="100">
                <div style="width:20px; height:150px;"></div>
            </td>
        </table>
    </td>
    <td style="background:lightblue;" data-expected-width="200">
        <div style="width:40px; height:150px;"></div>
    </td>
</table>
<script>
checkLayout('#outerTable');
</script>
